Fix image directory for recipe pages
Fix the problem of images not appearing in the github hosted website. This was done by changing the "../../" directory to be "../" instead. It works perfectly well in my local machine through
the Live Server extension of VS Code. After I'm done writing this commit I really hope the images turn up this time else I'm gonna have to think of another solution.